# perf(core): limit concurrent network requests